+++ benchmarks/verus/humaneval_from_dafny_humaneval/files/humaneval_060.rs
@@ -18,9 +18,9 @@ spec fn sum_from_one_to_n(n: int) -> int
 // </vc-helpers>
 
 // <vc-spec>
-fn sum_to_n(n: int) -> (result: int)
-    requires valid_input(n)
-    ensures result == sum_from_one_to_n(n)
+fn sum_to_n(n: i8) -> (result: i8)
+    requires valid_input(n as int)
+    ensures result as int == sum_from_one_to_n(n as int)
 // </vc-spec>
 // <vc-code>
 {
